requirements.txt
black==25.1.0
Jinja2==3.1.5
jsonpath==0.82.2
loguru==0.7.2
PyMySQL
pytest==7.4.2
PyYAML==6.0
requests==2.31.0
allure-pytest==2.8.24
colorlog==6.7.0
jsonschema==4.19.1
genson==1.2.2
click==8.1.3
emoji==2.2.0
click-help-colors==0.9.1
tenacity==8.2.3
ruamel.yaml==0.17.21
pydantic==2.10.5
jsonschema-extractor==2.0.2
rich==12.6.0
attrs==25.1.0
cattrs==24.1.2
translate==3.6.1
inflection==0.5.1
fastapi==0.110.0
uvicorn==0.28.0
wsproto==1.2.0
PyJWT==2.10.1